feat: add rnn trained in a remote gpu workspace
The model had to be severely decreased in accuracy to be able
to run in my machine. For this reason It didn't achieve the 3.5
loss expected out of it. This commit allows the model to have the
complexity it needs to lower the loss if it is trained on a
remote machine.